table modification

+++ database/migrations/2022_07_17_211350_create_breeds_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('breeds', function (Blueprint $table) {
             $table->id();
             $table->string('name', 200);
-            $table->unsignedBigInteger('type');
+            $table->unsignedSmallInteger('type');
             $table->timestamps();
         });
     }
